requirements.txt
pandas~=2.2.2
numpy~=1.26.4
mlxtend~=0.23.1
python-dotenv~=1.0.1

openai~=1.35.10
typing_extensions~=4.11.0
langgraph~=0.1.5
SQLAlchemy~=2.0.30
langsmith~=0.1.83
langchain~=0.2.6
langchain-anthropic~=0.1.19
langchain-community~=0.2.6
langchain-core~=0.2.0
langchain-experimental~=0.0.62
langchain-openai~=0.1.14
pytz~=2024.1
requests~=2.31.0
psycopg2-binary==2.9.9
pgvector~=0.2.5
fastapi~=0.111.0
pydantic~=2.7.1
starlette~=0.37.2
ipython~=8.24.0
aiosqlite~=0.20.0